requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-tavily>=0.2.0
python-dotenv>=1.0.0
# Vector store (FAISS — no pydantic dependency, Python 3.14 compatible)
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0


# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0

# Code quality
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0

# Documentation
markdown>=3.5.0
aiohttp>=3.9.0
